+++ Demo/Shared/Debug/ViewDebug.swift
@@ -1,0 +1,15 @@
+//
+//  ViewDebug.swift
+//  Demo
+//
+//  Created by Daniel Saidi on 2022-12-15.
+//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+final class ViewDebug {
+
+    /// Set this to true to enable view debugging.
+    static var isEnabled = false
+}